EP-57: Some minor fixes

+++ oneplan/oneplan/src/main/java/fit/biesp/oneplan/service/FriendService.java
@@ -1,6 +1,7 @@
 package fit.biesp.oneplan.service;
 
 import fit.biesp.oneplan.entity.FriendEntity;
+import fit.biesp.oneplan.exception.PersonNotFoundException;
 import fit.biesp.oneplan.model.FriendModel;
 import fit.biesp.oneplan.repository.FriendRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -11,25 +12,28 @@ public class FriendService {
     @Autowired
     private FriendRepository friendRepository;
 
-    public FriendModel addFriend(FriendModel friendModel){
+    public FriendModel addFriend(FriendModel friendModel) {
         return FriendModel.toModel(friendRepository.save(FriendModel.fromModel(friendModel)));
     }
 
-    public FriendModel getFriend(String nickname){
+    public FriendModel getFriend(String nickname) throws PersonNotFoundException {
         FriendEntity friend = friendRepository.findByNickname(nickname);
+        if (friend == null) throw new PersonNotFoundException("");
         return FriendModel.toModel(friend);
     }
 
-    public FriendModel updateFriend(FriendModel friendModel, String nickname){
+    public FriendModel updateFriend(FriendModel friendModel, String nickname) throws PersonNotFoundException {
         FriendEntity friend = friendRepository.findByNickname(nickname);
+        if (friend == null) throw new PersonNotFoundException("");
         friend.setNickname(friendModel.getNickname());
         friend.setEmail(friendModel.getEmail());
 
         return FriendModel.toModel(friendRepository.save(friend));
     }
 
-    public Long delete(String nickname){
+    public Long delete(String nickname) throws PersonNotFoundException {
         FriendEntity friend = friendRepository.findByNickname(nickname);
+        if (friend == null) throw new PersonNotFoundException("");
         Long id = friend.getId();
         friendRepository.deleteById(id);
         return id;
